% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateFunctions.R
\name{AddDays}
\alias{AddDays}
\title{AddDays}
\usage{
AddDays(DATE_VAL, ADD_N = NULL, DEBUG = TRUE)
}
\arguments{
\item{DATE_VAL}{The original date value that will be added to. Must be a date format}

\item{ADD_N}{Number of days to add to DATE_VAL}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
A single date value
}
\description{
This function will take a single date value and add N days to it.
}
\examples{
\donttest{

    AddDays(as.Date("2020-01-01"), 1)

}
}
\seealso{
Other Data Preparation: 
\code{\link{AddMonths}()},
\code{\link{AddWeeks}()},
\code{\link{CreateOutputDates}()},
\code{\link{FakeDataGenerator}()},
\code{\link{RemoveSpecialChars}()}
}
\author{
Abraham Mathew
}
\concept{Data Preparation}
