% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineeringFunctions.R
\name{AddEngineeredDates}
\alias{AddEngineeredDates}
\title{AddEngineeredDates}
\usage{
AddEngineeredDates(
  AGG_TS_DATA,
  DATE_COLUMN = "date",
  ADD_CHRISTMAS = TRUE,
  ADD_NEW_YEAR = TRUE,
  ADD_OTHER_HOLIDAY = TRUE,
  ADD_COVID_SHOCK = TRUE,
  DEBUG = TRUE
)
}
\arguments{
\item{AGG_TS_DATA}{The raw dataset that will be aggregated}

\item{DATE_COLUMN}{The column name containing date value}

\item{ADD_CHRISTMAS}{Add a feature to identify Christmas}

\item{ADD_NEW_YEAR}{Add a feature to identify New Years}

\item{ADD_OTHER_HOLIDAY}{Add a feature to identify Other Holidays}

\item{ADD_COVID_SHOCK}{Add a feature to identify the months where Covid was prominent}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
returns a data table with the aggregated data by date and any potential group variables
}
\description{
This function will take raw data and aggregate it by date
}
\seealso{
Other Feature Engineering: 
\code{\link{CreateTimeSeriesFeatures}()}
}
\author{
Abraham Mathew
}
\concept{Feature Engineering}
