% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrepFunctions.R
\name{AddMissingDates}
\alias{AddMissingDates}
\title{AddMissingDates}
\usage{
AddMissingDates(
  AGG_TS_DATA,
  DATE_COLUMN = "date",
  TARGET_COLUMN = "vals",
  TIME_FRAME = "day",
  IMPUTE_VALUE = 0.1,
  DEBUG = TRUE
)
}
\arguments{
\item{AGG_TS_DATA}{The raw dataset that will be aggregated}

\item{DATE_COLUMN}{The column name containing date value}

\item{TARGET_COLUMN}{The column name containing target value}

\item{TIME_FRAME}{The date values in the data (day, week, or month)}

\item{IMPUTE_VALUE}{The value to impute missing dates with}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
returns a data table with all missing dates imputed with a imputed value
}
\description{
This function will add missing dates in a time series
}
\seealso{
Other Data Preparation: 
\code{\link{AddDays}()},
\code{\link{AddMonths}()},
\code{\link{AddWeeks}()},
\code{\link{AggregateTimeSeriesData}()},
\code{\link{CheckForMissingDates}()},
\code{\link{CreateOutputDates}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FlattenLongData}()},
\code{\link{RemoveSpecialChars}()}
}
\author{
Abraham Mathew
}
\concept{Data Preparation}
