% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateFunctions.R
\name{AddMonths}
\alias{AddMonths}
\title{AddMonths}
\usage{
AddMonths(DATE_VAL, ADD_N = NULL, DEBUG = TRUE)
}
\arguments{
\item{DATE_VAL}{The original date value that will be added to. Must be a date format}

\item{ADD_N}{Number of months to add to DATE_VAL}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
A single date value
}
\description{
This function will take a single date value and add N months to it.
}
\examples{
\donttest{

    AddMonths(as.Date("2020-01-01"), 1)

}
}
\seealso{
Other Data Preparation: 
\code{\link{AddDays}()},
\code{\link{AddWeeks}()},
\code{\link{CreateOutputDates}()},
\code{\link{RemoveSpecialChars}()}
}
\author{
Abraham Mathew
}
\concept{Data Preparation}
