% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrepFunctions.R
\name{AggregateTimeSeriesData}
\alias{AggregateTimeSeriesData}
\title{AggregateTimeSeriesData}
\usage{
AggregateTimeSeriesData(
  RAW_TS_DATA,
  DATE_COLUMN = "SalesDate",
  TARGET_COLUMN = "TotalSales",
  OTHER_VARIABLES_TO_AGG = NULL,
  DIMENSION_COLUMN_NAMES = NULL,
  DEBUG = TRUE
)
}
\arguments{
\item{RAW_TS_DATA}{The raw dataset that will be aggregated}

\item{DATE_COLUMN}{The column name containing date value}

\item{TARGET_COLUMN}{The column name containing target value}

\item{OTHER_VARIABLES_TO_AGG}{Other columns that will be aggregated}

\item{DIMENSION_COLUMN_NAMES}{If value provided, aggregation will be done by this column}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
returns a data table with the aggregated data by date and any potential group variables
}
\description{
This function will take raw data and aggregate it by date
}
\seealso{
Other Data Preparation: 
\code{\link{AddDays}()},
\code{\link{AddMissingDates}()},
\code{\link{AddMonths}()},
\code{\link{AddWeeks}()},
\code{\link{CheckForMissingDates}()},
\code{\link{CreateOutputDates}()},
\code{\link{DoDataPartition}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FlattenLongData}()},
\code{\link{RemoveSpecialChars}()}
}
\author{
Abraham Mathew
}
\concept{Data Preparation}
