% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateFunctions.R
\name{CreateOutputDates}
\alias{CreateOutputDates}
\title{CreateOutputDates}
\usage{
CreateOutputDates(
  FULL_DATA,
  DATE_COLUMN = "date_ymd",
  DATA_TYPE_COLUMN = "DataType",
  FORECAST_HORIZON = 3,
  TIME_FRAME = "month",
  EVAL_MODE = TRUE,
  DEBUG = TRUE
)
}
\arguments{
\item{FULL_DATA}{The full data set that will be used in the forecasting models}

\item{DATE_COLUMN}{The column name of the date variable}

\item{DATA_TYPE_COLUMN}{The column name of the data type variable}

\item{FORECAST_HORIZON}{Forecast horizon for the prediction}

\item{TIME_FRAME}{Is the forecast data daily, weekly, or monthly data}

\item{EVAL_MODE}{If TRUE, only perform train and test on the actual data}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
A sequence of dates for the output from the predictive model
}
\description{
This function will collect date values for the output of the forecasting output
}
\examples{

\donttest{

     library(data.table)
     
     FULL_DATA <- data.table::data.table(date_ymd = seq.Date(as.Date("2010-01-01"), 
                                                             length.out = 12, by = "month"),
                                         DataType = c(rep("TRAIN",9), rep("TEST",3)))

      CreateOutputDates(FULL_DATA,
                                DATE_COLUMN = "date_ymd",
                                DATA_TYPE_COLUMN = "DataType",
                                FORECAST_HORIZON = 3,
                                TIME_FRAME = "month",
                                EVAL_MODE = TRUE,
                                DEBUG = TRUE)
                           

}

}
\seealso{
Other Data Preparation: 
\code{\link{AddDays}()},
\code{\link{AddMissingDates}()},
\code{\link{AddMonths}()},
\code{\link{AddWeeks}()},
\code{\link{AggregateTimeSeriesData}()},
\code{\link{CheckForMissingDates}()},
\code{\link{DoDataPartition}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FlattenLongData}()},
\code{\link{RemoveSpecialChars}()}
}
\author{
Abraham Mathew
}
\concept{Data Preparation}
