% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{CreateProjectFiles}
\alias{CreateProjectFiles}
\title{CreateProjectFiles}
\usage{
CreateProjectFiles(
  PROJECT_PATH = rstudioapi::getActiveProject(),
  FOLDERS_TO_CREATE = c("data", "docs", "figs", "logs", "output", "queries", "R",
    "tests"),
  OPEN_NEW_SESSION = TRUE,
  DEBUG = TRUE
)
}
\arguments{
\item{PROJECT_PATH}{The main project directory}

\item{FOLDERS_TO_CREATE}{The folders that will be created in the project path}

\item{OPEN_NEW_SESSION}{If TRUE, a new session will be generated}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
Creates a project and folders in a directory if a RStudio project is not present
}
\description{
Check a directory and create a project and folders if they don't exist
}
\seealso{
Other Project Preparation: 
\code{\link{PackageCheck}()}
}
\author{
Abraham Mathew
}
\concept{Project Preparation}
