% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineeringFunctions.R
\name{CreateTimeSeriesFeatures}
\alias{CreateTimeSeriesFeatures}
\title{CreateTimeSeriesFeatures}
\usage{
CreateTimeSeriesFeatures(
  AGG_TS_DATA,
  DATE_COLUMN = "SalesDate",
  WHICH_FEATURES = NULL,
  BY_WHICH_GROUP = "Currency",
  TIME_UNIT = "day",
  LAG_VALS = c(1:10),
  MA_VALS = c(5:10),
  SIMPLE_IMPUTE = TRUE,
  DEBUG = TRUE
)
}
\arguments{
\item{AGG_TS_DATA}{The raw dataset that will be aggregated}

\item{DATE_COLUMN}{The column name containing date value}

\item{WHICH_FEATURES}{The column names of the original features that we will use}

\item{BY_WHICH_GROUP}{If value provided, features will be generated by this column}

\item{TIME_UNIT}{The time frame that is represented in the dataset  (day, week, or month)}

\item{LAG_VALS}{The numbef of values of the time unit that will be used to create the lagged features}

\item{MA_VALS}{The numbef of values of the time unit that will be used to create the rolling average features}

\item{SIMPLE_IMPUTE}{If TRUE, missing values will be imputed with 0.1}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
returns the originak data table with all the new features appended to the dataset
}
\description{
This function will take aggregated time series data and generate various features
}
\seealso{
Other Feature Engineering: 
\code{\link{AddEngineeredDates}()}
}
\author{
Abraham Mathew
}
\concept{Feature Engineering}
