% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelingFunctions.R
\name{DoDataPartition}
\alias{DoDataPartition}
\title{DoDataPartition}
\usage{
DoDataPartition(
  DATA_DT,
  DATE_COLUMN = "SalesDate",
  TARGET_COLUMN = "Sales",
  WHICH_FEATURES = NULL,
  TRAIN_RATIO = 0.8,
  PARTITION_TYPE = "time",
  DEBUG = TRUE
)
}
\arguments{
\item{DATA_DT}{A data.table which contains the main dataset}

\item{DATE_COLUMN}{The variable which contains date values}

\item{TARGET_COLUMN}{The variable which contains the target variable}

\item{WHICH_FEATURES}{The variables that will need to be in the final dataset}

\item{TRAIN_RATIO}{Percentage of rows that will be in the dataset}

\item{PARTITION_TYPE}{Partition by time or randomly}

\item{DEBUG}{If TRUE, the function will be run in debug mode}
}
\description{
This function will take one dataset and return data for training and testing
}
\seealso{
Other Data Preparation: 
\code{\link{AddDays}()},
\code{\link{AddMissingDates}()},
\code{\link{AddMonths}()},
\code{\link{AddWeeks}()},
\code{\link{AggregateTimeSeriesData}()},
\code{\link{CheckForMissingDates}()},
\code{\link{CreateOutputDates}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FlattenLongData}()},
\code{\link{RemoveSpecialChars}()}
}
\author{
Abraham Mathew
}
\concept{Data Preparation}
