% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelingFunctions.R
\name{DoOneHotEncoding}
\alias{DoOneHotEncoding}
\title{DoOneHotEncoding}
\usage{
DoOneHotEncoding(
  MODEL_DATA,
  COLS_TO_ENCODE,
  DROP_ORIGINAL_COLS = FALSE,
  DROP_UNUSED_LEVELS = FALSE,
  DEBUG = TRUE
)
}
\arguments{
\item{MODEL_DATA}{The processed dataset that will be used in the predictive model}

\item{COLS_TO_ENCODE}{The column names containing character values that need to be encoded}

\item{DROP_ORIGINAL_COLS}{Remove the original character variable}

\item{DROP_UNUSED_LEVELS}{Drop any levels that are missing (should be FALSE)}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
returns the originak data table with all the new features appended to the dataset
}
\description{
This function will take specified column names and append them as one hot encoded variables
}
\seealso{
Other Modeling Functions: 
\code{\link{RunAllModels}()},
\code{\link{mae}()},
\code{\link{rmse}()}
}
\author{
Abraham Mathew
}
\concept{Modeling Functions}
