% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FakeDataGenerator.R
\name{FakeDataGenerator}
\alias{FakeDataGenerator}
\title{FakeDataGenerator}
\usage{
FakeDataGenerator(
  NUM_ROWS = 10000,
  DATE_COLS = 1,
  NUMERIC_COLS = 3,
  CHARACTER_COLS = 1,
  LOGICAL_COLS = 0,
  FACTOR_COLS = 2,
  DEBUG = TRUE
)
}
\arguments{
\item{NUM_ROWS}{Number of records}

\item{DATE_COLS}{Number of date columns}

\item{NUMERIC_COLS}{Number of numeric columns}

\item{CHARACTER_COLS}{Number of character columns}

\item{LOGICAL_COLS}{Number of logical columns}

\item{FACTOR_COLS}{Number of factor columns}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
A data table with some fake data
}
\description{
Create fake data for examples
}
\examples{

\dontrun{

data <- GoodeR::FakeDataGenerator(NUM_ROWS = 10000,
                           DATE_COLS = 1,
                           NUMERIC_COLS = 3,
                           CHARACTER_COLS = 1,
                           LOGICAL_COLS = 0,
                           FACTOR_COLS = 2,
                           DEBUG = TRUE)
                           
}

}
\seealso{
Other Data Preparation: 
\code{\link{AddDays}()},
\code{\link{AddMissingDates}()},
\code{\link{AddMonths}()},
\code{\link{AddWeeks}()},
\code{\link{AggregateTimeSeriesData}()},
\code{\link{CheckForMissingDates}()},
\code{\link{CreateOutputDates}()},
\code{\link{DoDataPartition}()},
\code{\link{FlattenLongData}()},
\code{\link{RemoveSpecialChars}()}
}
\author{
Abraham Mathew
}
\concept{Data Preparation}
