% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrepFunctions.R
\name{FlattenLongData}
\alias{FlattenLongData}
\title{FlattenLongData}
\usage{
FlattenLongData(
  WHICH_LONG_DATA,
  GROUP_BY_VAR = c("Date_new"),
  NEW_COLUMN_VAR = c("Currency"),
  VALUE_VAR = c("Amount_new"),
  AGG_METHOD = c("sum"),
  DEBUG = TRUE
)
}
\arguments{
\item{WHICH_LONG_DATA}{A data.table which contains data in long format}

\item{GROUP_BY_VAR}{The variables that will comprise the grouping columns}

\item{NEW_COLUMN_VAR}{The variables that will be added as new columns}

\item{VALUE_VAR}{The numeric variable that will be aggregated in the new columns}

\item{AGG_METHOD}{Aggregation method that will be used on each element in the ValueVar}

\item{DEBUG}{If TRUE, the function will be run in debug mode}
}
\value{
Returns a data table with the flattened dataset
}
\description{
This function will take data in long format and flatten it
}
\seealso{
Other Data Preparation: 
\code{\link{AddDays}()},
\code{\link{AddMissingDates}()},
\code{\link{AddMonths}()},
\code{\link{AddWeeks}()},
\code{\link{AggregateTimeSeriesData}()},
\code{\link{CheckForMissingDates}()},
\code{\link{CreateOutputDates}()},
\code{\link{DoDataPartition}()},
\code{\link{FakeDataGenerator}()},
\code{\link{RemoveSpecialChars}()}
}
\author{
Abraham Mathew
}
\concept{Data Preparation}
