% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImportFunctions.R
\name{GetCsvData}
\alias{GetCsvData}
\title{GetCsvData}
\usage{
GetCsvData(
  PROJECT_PATH = rstudioapi::getActiveProject(),
  WHICH_DATA_FOLDER = "processed",
  DEBUG = TRUE
)
}
\arguments{
\item{PROJECT_PATH}{The main project directory}

\item{WHICH_DATA_FOLDER}{The name of the data sub directory from where data will be pulled}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
Returns all the csv files in the data folder within a project as a data.table
}
\description{
This function will grab data from all csv files within a directory
}
\seealso{
Other Data Import: 
\code{\link{GetSqlData}()},
\code{\link{GetSqlQuery}()}
}
\author{
Abraham Mathew
}
\concept{Data Import}
