% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImportFunctions.R
\name{GetSqlData}
\alias{GetSqlData}
\title{GetSqlData}
\usage{
GetSqlData(
  DRIVER = "SQL Server",
  SERVER = "sdl02-vm161",
  DB_NAME = "OpsDW",
  PROJECT_PATH = NULL,
  DEBUG = TRUE
)
}
\arguments{
\item{DRIVER}{Name of the SQL Platform}

\item{SERVER}{Name of the server}

\item{DB_NAME}{Name of the database}

\item{PROJECT_PATH}{path to the project directory}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
Returns the results of the sql query
}
\description{
This function will grab data from a SQL database using a SQL file
}
\seealso{
Other Data Import: 
\code{\link{GetCsvData}()},
\code{\link{GetSqlQuery}()}
}
\author{
Abraham Mathew
}
\concept{Data Import}
