% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImportFunctions.R
\name{GetSqlQuery}
\alias{GetSqlQuery}
\title{GetSqlQuery}
\usage{
GetSqlQuery(PROJECT_PATH = rstudioapi::getActiveProject(), DEBUG = TRUE)
}
\arguments{
\item{PROJECT_PATH}{The main project directory}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
Returns the sql query as a srring
}
\description{
This function will grab the conrents if a SQL file as a string
}
\seealso{
Other Data Import: 
\code{\link{GetCsvData}()},
\code{\link{GetSqlData}()}
}
\author{
Abraham Mathew
}
\concept{Data Import}
