% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoodeR-package.R
\docType{package}
\name{GoodeR-package}
\alias{GoodeR}
\alias{GoodeR-package}
\title{GoodeR: Automate Common Data Science Tasks}
\description{
R package for the automation of machine learning, forecasting, feature engineering, model evaluation, model interpretation, and other tasks. Built using data.table for all tabular data-related tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abmathewks/GoodeR}
  \item Report bugs at \url{https://github.com/abmathewks/GoodeR/issues}
}

}
\author{
\strong{Maintainer}: Abraham Mathew \email{mathewanalytics@gmail.com} [contributor]

}
\keyword{internal}
