% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelingFunctions.R
\name{mae}
\alias{mae}
\title{mae}
\usage{
mae(actual, predicted)
}
\arguments{
\item{actual}{Actual data from the test data}

\item{predicted}{Predicted values from the predictive model}
}
\value{
A single numeric value
}
\description{
Caculate the mae metric from model predictions
}
\seealso{
Other Modeling Functions: 
\code{\link{DoOneHotEncoding}()},
\code{\link{RunAllModels}()},
\code{\link{rmse}()}
}
\author{
Abraham Mathew
}
\concept{Modeling Functions}
