% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{PackageCheck}
\alias{PackageCheck}
\title{PackageCheck}
\usage{
PackageCheck(USE_THESE_PACKAGES, DEBUG = TRUE)
}
\arguments{
\item{USE_THESE_PACKAGES}{A vector of strings which describe the packages that will be used during the project}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
Installs any missing packages and loads of them into the current workspace
}
\description{
Check if a vector of packages is installed. If they are not, install and load them
}
\examples{

\donttest{

     GoodeR::PackageCheck(c("ggplot2","data.table","lubridate","rms"))

}

}
\seealso{
Other Project Preparation: 
\code{\link{CreateProjectFiles}()}
}
\author{
Abraham Mathew
}
\concept{Project Preparation}
