% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelingFunctions.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{Actual data from the test data}

\item{predicted}{Predicted values from the predictive model}
}
\value{
A single numeric value
}
\description{
Caculates the rmse metric from model predictions
}
\seealso{
Other Modeling Functions: 
\code{\link{DoOneHotEncoding}()},
\code{\link{RunAllModels}()},
\code{\link{mae}()}
}
\author{
Abraham Mathew
}
\concept{Modeling Functions}
