% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateFunctions.R
\name{CreateOutputDates}
\alias{CreateOutputDates}
\title{CreateOutputDates}
\usage{
CreateOutputDates(
  FULL_DATA,
  DATE_COLUMN = "SalesDate",
  DATA_TYPE_COLUMN = "DataType",
  FORECAST_HORIZON = 30,
  TIME_FRAME = "day",
  EVAL_MODE = TRUE,
  DEBUG = TRUE
)
}
\arguments{
\item{FULL_DATA}{The full data set that will be used in the forecasting models}

\item{DATE_COLUMN}{The column name of the date variable}

\item{DATA_TYPE_COLUMN}{The column name of the data type variable}

\item{FORECAST_HORIZON}{Forecast horizon for the prediction}

\item{TIME_FRAME}{Is the forecast data daily, weekly, or monthly data}

\item{EVAL_MODE}{If TRUE, only perform train and test on the actual data}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\description{
This function will collect date values for the output of the forecasting output
}
\seealso{
Other Data Preparation: 
\code{\link{AddDays}()},
\code{\link{AddMissingDates}()},
\code{\link{AddMonths}()},
\code{\link{AddWeeks}()},
\code{\link{AggregateTimeSeriesData}()},
\code{\link{CheckForMissingDates}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FlattenLongData}()},
\code{\link{RemoveSpecialChars}()}
}
\author{
Abraham Mathew
}
\concept{Data Preparation}
