% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataImportFunctions.R
\name{GetSqlQuery}
\alias{GetSqlQuery}
\title{GetSqlQuery}
\usage{
GetSqlQuery(SQL_QUERY_FILEPATH, DEBUG = TRUE)
}
\arguments{
\item{SQL_QUERY_FILEPATH}{Path to the .SQL file that needs to be imported}

\item{DEBUG}{If TRUE, the function will run in debug mode}
}
\value{
returns the sql query as a srring
}
\description{
This function will grab the conrents if a SQL file as a string
}
\seealso{
Other Data Import: 
\code{\link{GetCsvData}()},
\code{\link{GetSqlData}()}
}
\author{
Abraham Mathew
}
\concept{Data Import}
